Add files via upload
Hi this is BOB the Biped, the 3d file is not mine but from http://www.thingiverse.com/thing:43708

The code was originally made by me, feel free to use it, although BOB does not bite, it is your responsibility that this project in any way does not harm or inflict damage of any sort. 

The code will make BOB walk and jump on every restart of the code, I suggest using at least 1000mA (1A) PSU for this project as the servos draw quite a few amps. 

Comments and suggestions are always welcomed :-)